MGMT-8392: Feature support level indications in cluster details
