+++ src/ocm/components/featureSupportLevels/index.ts
@@ -1,3 +1,3 @@
 export { default as FeatureSupportLevelProvider } from './FeatureSupportLevelProvider';
-export { default as ReviewClusterFeatureSupportLevels } from './ReviewClusterFeatureSupportLevels';
+export { default as ClusterFeatureSupportLevelsDetailItem } from './ReviewClusterFeatureSupportLevels';
 export * from './utils';
